Code refactoring by Thomas Harning (Bug [#767]) Removes use of deprecated implicit `arg`.. partly through the next major fix/enhancement.
Moved to a tail-recursive method of iterating through the returnvalues.

Basically coxpcall created a coroutine then executes performResume.

performResume in turn calls handleReturnValue which basically filters out the `status` value while not using 'arg' or
select.  That in turn checks for error condition and acts appropriately.  If the coro is suspended, then it jumps to
performResume...

It's kind of loopy, but its use of tail recursion prevents stack overflow and the use of Lua 5.1's var-arg processing
removes the requirement for allocating tables.